+++ example/android/app/src/main/kotlin/com/savelee/flutter_app/MainActivity.kt
@@ -1,0 +1,6 @@
+package com.savelee.flutter_app
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity: FlutterActivity() {
+}