+++ example/codelab-steps/chat-v1.dart
@@ -1,0 +1,210 @@
+import 'package:flutter/material.dart';
+
+import 'dart:async';
+import 'package:flutter/services.dart';
+import 'package:rxdart/rxdart.dart';
+import 'package:sound_stream/sound_stream.dart';
+
+//TODO import Dialogflow
+
+class Chat extends StatefulWidget {
+  Chat({Key key}) : super(key: key);
+
+  @override
+  _ChatState createState() => _ChatState();
+}
+
+class _ChatState extends State<Chat> {
+  final List<ChatMessage> _messages = <ChatMessage>[];
+  final TextEditingController _textController = new TextEditingController();
+
+  bool _isRecording = false;
+
+  RecorderStream _recorder = RecorderStream();
+  StreamSubscription _recorderStatus;
+  StreamSubscription<List<int>> _audioStreamSubscription;
+  BehaviorSubject<List<int>> _audioStream;
+
+  // TODO DialogflowGrpc class instance
+
+  @override
+  void initState() {
+    super.initState();
+    initPlugin();
+  }
+
+  @override
+  void dispose() {
+    _recorderStatus?.cancel();
+    _audioStreamSubscription?.cancel();
+    super.dispose();
+  }
+
+  // Platform messages are asynchronous, so we initialize in an async method.
+  Future<void> initPlugin() async {
+    _recorderStatus = _recorder.status.listen((status) {
+      if (mounted)
+        setState(() {
+          _isRecording = status == SoundStreamStatus.Playing;
+        });
+    });
+
+    await Future.wait([
+      _recorder.initialize()
+    ]);
+
+    // TODO Get a Service account
+    // TODO Create a DialogflowGrpc Instance
+  }
+
+  void stopStream() async {
+    await _recorder.stop();
+    await _audioStreamSubscription?.cancel();
+    await _audioStream?.close();
+  }
+
+  void handleSubmitted(text) async {
+    print(text);
+    _textController.clear();
+
+    //TODO Dialogflow Code
+  }
+
+  void handleStream() async {
+    _recorder.start();
+
+    _audioStream = BehaviorSubject<List<int>>();
+    _audioStreamSubscription = _recorder.audioStream.listen((data) {
+      print(data);
+      _audioStream.add(data);
+    });
+
+    // TODO Create and audio InputConfig
+    //  See: https://cloud.google.com/dialogflow/es/docs/reference/rpc/google.cloud.dialogflow.v2#google.cloud.dialogflow.v2.InputAudioConfig
+
+    // TODO Make the streamingDetectIntent call, with the InputConfig and the audioStream
+
+    // TODO Get the transcript and detectedIntent and show on screen
+
+  }
+
+  // The chat interface
+  //
+  //------------------------------------------------------------------------------------
+  @override
+  Widget build(BuildContext context) {
+    return Column(children: <Widget>[
+      Flexible(
+          child: ListView.builder(
+            padding: EdgeInsets.all(8.0),
+            reverse: true,
+            itemBuilder: (_, int index) => _messages[index],
+            itemCount: _messages.length,
+          )),
+      Divider(height: 1.0),
+      Container(
+          decoration: BoxDecoration(color: Theme.of(context).cardColor),
+          child: IconTheme(
+            data: IconThemeData(color: Theme.of(context).accentColor),
+            child: Container(
+              margin: const EdgeInsets.symmetric(horizontal: 8.0),
+              child: Row(
+                children: <Widget>[
+                  Flexible(
+                    child: TextField(
+                      controller: _textController,
+                      onSubmitted: handleSubmitted,
+                      decoration: InputDecoration.collapsed(hintText: "Send a message"),
+                    ),
+                  ),
+                  Container(
+                    margin: EdgeInsets.symmetric(horizontal: 4.0),
+                    child: IconButton(
+                      icon: Icon(Icons.send),
+                      onPressed: () => handleSubmitted(_textController.text),
+                    ),
+                  ),
+                  IconButton(
+                    iconSize: 30.0,
+                    icon: Icon(_isRecording ? Icons.mic_off : Icons.mic),
+                    onPressed: _isRecording ? stopStream : handleStream,
+                  ),
+                ],
+              ),
+            ),
+          )
+      ),
+    ]);
+  }
+}
+
+
+//------------------------------------------------------------------------------------
+// The chat message balloon
+//
+//------------------------------------------------------------------------------------
+class ChatMessage extends StatelessWidget {
+  ChatMessage({this.text, this.name, this.type});
+
+  final String text;
+  final String name;
+  final bool type;
+
+  List<Widget> otherMessage(context) {
+    return <Widget>[
+      new Container(
+        margin: const EdgeInsets.only(right: 16.0),
+        child: new CircleAvatar(child: new Text('B')),
+      ),
+      new Expanded(
+        child: new Column(
+          crossAxisAlignment: CrossAxisAlignment.start,
+          children: <Widget>[
+            new Text(this.name,
+                style: new TextStyle(fontWeight: FontWeight.bold)),
+            new Container(
+              margin: const EdgeInsets.only(top: 5.0),
+              child: new Text(text),
+            ),
+          ],
+        ),
+      ),
+    ];
+  }
+
+  List<Widget> myMessage(context) {
+    return <Widget>[
+      new Expanded(
+        child: new Column(
+          crossAxisAlignment: CrossAxisAlignment.end,
+          children: <Widget>[
+            new Text(this.name, style: Theme.of(context).textTheme.subhead),
+            new Container(
+              margin: const EdgeInsets.only(top: 5.0),
+              child: new Text(text),
+            ),
+          ],
+        ),
+      ),
+      new Container(
+        margin: const EdgeInsets.only(left: 16.0),
+        child: new CircleAvatar(
+            child: new Text(
+              this.name[0],
+              style: new TextStyle(fontWeight: FontWeight.bold),
+            )),
+      ),
+    ];
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return new Container(
+      margin: const EdgeInsets.symmetric(vertical: 10.0),
+      child: new Row(
+        crossAxisAlignment: CrossAxisAlignment.start,
+        children: this.type ? myMessage(context) : otherMessage(context),
+      ),
+    );
+  }
+}